Access both onPress and onLongPress
So in my case I have a card. When I press it (onPress) I need to activate something (send something to server), and when I'm pressing for some seconds (onLongPress) I want to open the menu. I couldn't do it without changing the code on MenuTrigger.js. If you find this useful you can add it. If somehow I can achieve the same thing without changing the repo code I would be glad to know how.